Add Safari hacks to minimize reflow issues
Attempt to Fix #53 with a very bad workaround.

At least we encapsulate the workaround in a safari "targeting" media
query so that Firefox/Chrome/IE/Edge are not affected by the workaround.

This is to minimize the effects of a weird bug where Safari randomly
applies the horizontal border spacing (either set with CELLSPACING
attribute or border-spacing property) on the right side of the table
when redrawing/reflowing.

Also change a lot of display: none/block to visibility: hidden/visible
when used in ":hover" conditions: maybe the browser needs less reflow
when the item is already displayed but invisible.