First implementation of image preloading
Show some loading animation while preloading images.
For placeholder we even try to create them on the client side using SVG: it doesn't make sense to show a local placeholder while waiting a remote placeholder, but at this time we need the remote one and we want to test local generation.