introduce placeholderText() function to define the text for image placeholders
ko.bindingHandlers.wysiwygSrc.placeholderText(width, height) can be overridden and may be called with only one dimension when the placeholder is relative to a space that will resize instead of cropping the image.

Also, this "text" is passed as a parameter to the backend call generating the placeholder, so that the backend can write it on the image.