fileupload: explicit width/height risizing + remove obsolete option
by default our fileupload resized large images client side to a maximum resolution of 1920x1080. I added this options to the default option object to make it clear it happens.

Then removed the "disableImageResize" for Android/Opera as it was a compatibility issue with old browsers that we probably don't support anymore for other runtime requirements.